% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotSBS.R
\name{NotSBS_not}
\alias{NotSBS_not}
\title{Change Point Detection via NotSBS}
\usage{
NotSBS_not(
  F_hat,
  m = NULL,
  beta = NULL,
  trim = NULL,
  threshold = NULL,
  method = c("fixed", "oracle", "none"),
  V_shap = NULL,
  lbd = NULL
)
}
\arguments{
\item{F_hat}{A numeric matrix of signals where each column corresponds to a time point.}

\item{m}{For the oracle method, an integer specifying the expected number of change points.
Ignored if \code{method} is not \code{"oracle"}. Note: if \code{m} is not provided,
internal logic may adjust it based on the supplied \code{type} (if available).}

\item{beta}{A numeric value intended to inform the seeding of intervals; currently not actively used
in the interval selection.}

\item{trim}{A positive numeric value specifying the trimming length; only intervals with length greater
than \code{2 * trim} will be considered.}

\item{threshold}{A numeric threshold for the fixed method. Must be supplied if \code{method = "fixed"}.}

\item{method}{A character string specifying the threshold selection method. Choices are \code{"fixed"},
\code{"oracle"} and \code{"none"}.}

\item{V_shap}{(Optional) A character string indicating the form of the variance estimator to use.
Choices are \code{"diag"}, \code{"full"}, or \code{"non"}. If not supplied, no variance
correction is applied.}

\item{lbd}{A numeric value specifying the minimum length for the seeded intervals (often set as a multiple
of \code{trim}).}
}
\value{
A data frame containing the detected change points along with associated statistics. For the oracle
method, additional columns \code{selected_threshold}, \code{next_highest_threshold}, and
\code{no_change_threshold} are appended.
}
\description{
This function performs change point detection on a time series by utilising a matrix of estimated signals,
\code{F_hat}, and applying either a fixed or an oracle method for threshold selection.
}
\examples{
\dontrun{
# Assuming F_hat is a matrix with time series data as columns:
results_fixed <- NotSBS_not(F_hat = F_hat, m = 3, type = c(4,1,2), beta = 0.5, trim = 10,
                            threshold = 5, method = "fixed",
                            V_shap = "diag", lbd = 20)

results_oracle <- NotSBS_not(F_hat = F_hat, m = 3, beta = 0.5, trim = 10,
                             method = "oracle",
                             V_shap = "diag", lbd = 20)
}
}
